Fixes a numerical accuracy issue in the pitch search
The energy "sliding window" was using double accumulation with
float multiplications. This forces the multiplications to be
double as well.

+++ silk/float/energy_FLP.c
@@ -41,18 +41,18 @@ double silk_energy_FLP(
     double   result;
 
     /* 4x unrolled loop */
-    result = 0.0f;
+    result = 0.0;
     dataSize4 = dataSize & 0xFFFC;
     for( i = 0; i < dataSize4; i += 4 ) {
-        result += data[ i + 0 ] * data[ i + 0 ] +
-                  data[ i + 1 ] * data[ i + 1 ] +
-                  data[ i + 2 ] * data[ i + 2 ] +
-                  data[ i + 3 ] * data[ i + 3 ];
+        result += data[ i + 0 ] * (double)data[ i + 0 ] +
+                  data[ i + 1 ] * (double)data[ i + 1 ] +
+                  data[ i + 2 ] * (double)data[ i + 2 ] +
+                  data[ i + 3 ] * (double)data[ i + 3 ];
     }
 
     /* add any remaining products */
     for( ; i < dataSize; i++ ) {
-        result += data[ i ] * data[ i ];
+        result += data[ i ] * (double)data[ i ];
     }
 
     silk_assert( result >= 0.0 );
